user-controls: Fix sense of label for disabling app restrictions
Currently, if the app filter is disabled for a user (i.e. no apps are
filtered), the UI says:

Application Suitability: All Ages

which is the wrong way round. Disabling app restrictions means apps
which are not suitable for all ages may be browsed/installed.

Change the setting option from ‘All Ages’ to ‘Unrestricted’, which
is the right sense and fits in with the wording of the setting’s
description.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Fixes: #68